factor out process_body() for easier testing
