+++ modules/ngx_http_cbor_ic_modules/base32.c
