+++ modules/ngx_http_cbor_ic_modules/process_body.h
@@ -1,0 +1,9 @@
+#include "ngx_http_cbor_req_ic_module.h"
+
+static void nullify_str(ngx_str_t *s)
+{
+    s->data = NULL;
+    s->len = 0;
+}
+
+void process_body(buf_t b, ngx_http_cbor_req_ic_ctx_t *ctx);
