+++ modules/ngx_http_cbor_ic_modules/process_body_test.c
@@ -1,0 +1,54 @@
+#ifdef __IC_RUN_TESTS
+#include "process_body.h"
+#include "identifier.h"
+#include <string.h>
+#include <assert.h>
+
+unsigned char cbor_request_call[146] = {0xd9, 0xd9, 0xf7, 0xa1, 0x67, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0xa7, 0x6c, 0x72, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x5f, 0x74, 0x79, 0x70, 0x65, 0x64, 0x63, 0x61, 0x6c, 0x6c, 0x65, 0x6e, 0x6f, 0x6e, 0x63, 0x65, 0x50, 0x2e, 0x6d, 0x86, 0x08, 0x03, 0xdb, 0xb3, 0x8c, 0x00, 0xd9, 0xdd, 0xa6, 0xce, 0xd8, 0x6f, 0x99, 0x6e, 0x69, 0x6e, 0x67, 0x72, 0x65, 0x73, 0x73, 0x5f, 0x65, 0x78, 0x70, 0x69, 0x72, 0x79, 0x1b, 0x17, 0x3d, 0x0b, 0x71, 0x24, 0x53, 0xb0, 0x98, 0x66, 0x73, 0x65, 0x6e, 0x64, 0x65, 0x72, 0x41, 0x04, 0x6b, 0x63, 0x61, 0x6e, 0x69, 0x73, 0x74, 0x65, 0x72, 0x5f, 0x69, 0x64, 0x4a, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x01, 0x01, 0x6b, 0x6d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x63, 0x73, 0x65, 0x74, 0x63, 0x61, 0x72, 0x67, 0x4f, 0x44, 0x49, 0x44, 0x4c, 0x00, 0x01, 0x78, 0x06, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00};
+unsigned char cbor_request_query[118] = {0xd9, 0xd9, 0xf7, 0xa1, 0x67, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0xa6, 0x6c, 0x72, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x5f, 0x74, 0x79, 0x70, 0x65, 0x65, 0x71, 0x75, 0x65, 0x72, 0x79, 0x6e, 0x69, 0x6e, 0x67, 0x72, 0x65, 0x73, 0x73, 0x5f, 0x65, 0x78, 0x70, 0x69, 0x72, 0x79, 0x1b, 0x17, 0x3d, 0x0c, 0x15, 0x71, 0x7b, 0xae, 0x30, 0x66, 0x73, 0x65, 0x6e, 0x64, 0x65, 0x72, 0x41, 0x04, 0x6b, 0x63, 0x61, 0x6e, 0x69, 0x73, 0x74, 0x65, 0x72, 0x5f, 0x69, 0x64, 0x4a, 0x00, 0x00, 0x00, 0x00, 0x01, 0x60, 0x08, 0xaa, 0x01, 0x01, 0x6b, 0x6d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x65, 0x62, 0x75, 0x74, 0x74, 0x73, 0x63, 0x61, 0x72, 0x67, 0x47, 0x44, 0x49, 0x44, 0x4c, 0x00, 0x01, 0x7f};
+
+ngx_http_cbor_req_ic_ctx_t *test_mk_ctx()
+{
+    ngx_http_cbor_req_ic_ctx_t *ctx = malloc(sizeof(ngx_http_cbor_req_ic_ctx_t));
+    assert(ctx != NULL);
+
+    nullify_str(&ctx->method_name);
+    nullify_str(&ctx->request_type);
+
+    ctx->canister_id.data = malloc(64);
+    ctx->sender.data = malloc(64);
+
+    ZERO(ctx->canister_id.data);
+    ZERO(ctx->sender.data);
+
+    assert(ctx->canister_id.data != NULL);
+    assert(ctx->sender.data != NULL);
+
+    return ctx;
+}
+
+int main()
+{
+    buf_t b;
+    b.start = &cbor_request_call[0];
+    b.end = &cbor_request_call[146];
+
+    ngx_http_cbor_req_ic_ctx_t *ctx = test_mk_ctx();
+
+    process_body(b, ctx);
+    assert(strncmp(ctx->canister_id.data, "rrkah-fqaaa-aaaaa-aaaaq-cai", ctx->canister_id.len) == 0);
+    assert(strncmp(ctx->sender.data, "2vxsx-fae", ctx->sender.len) == 0);
+    assert(strncmp(ctx->method_name.data, "set", ctx->method_name.len) == 0);
+    assert(strncmp(ctx->request_type.data, "call", ctx->request_type.len) == 0);
+
+    b.start = &cbor_request_query[0];
+    b.end = &cbor_request_query[118];
+
+    ctx = test_mk_ctx();
+    process_body(b, ctx);
+    assert(strncmp(ctx->canister_id.data, "z2rt2-eaaaa-aaaal-abcva-cai", ctx->canister_id.len) == 0);
+    assert(strncmp(ctx->sender.data, "2vxsx-fae", ctx->sender.len) == 0);
+    assert(strncmp(ctx->method_name.data, "butts", ctx->method_name.len) == 0);
+    assert(strncmp(ctx->request_type.data, "query", ctx->request_type.len) == 0);
+}
+#endif
